+++ libft/ft_strlcpy.c
@@ -1,20 +1,21 @@
+#include <stddef.h>
 #include <stdlib.h>
 
 size_t	ft_strlcpy(char *dst, const char *src, size_t dstsize)
 {
-	unsigned long int i;
+	size_t	i;
 
-	i=0;
-	if (!(dstsize==0))
+	i = 0;
+	if (!(dstsize == 0))
 	{
-		while (i<(dstsize-1))
+		while (i < (dstsize - 1))
 		{
-			dst[i]=src[i];
+			dst[i] = src[i];
 			i++;
 		}
 	}
-	i=0;
-	while(src[i]!='\0')
+	i = 0;
+	while (src[i] != '\0')
 		i++;
 	return (i);
 }
